Create CI (#24) PR: #28
Add GitHub Actions CI

check.yml (on push)
- checks style with stylua
- static analysis with luacheck
- runs tests with busted

publish.yml (on tag push)
- runs check.yml
- checks version
- uploads to luarocks

- Change 23.0 to 23

Lua 5.3 changed the behaviour of converting a number to a string.

Prior to 5.3:
tostring(10.0) -- "10"

After 5.3:
tostring(10.0) -- "10.0"

http://www.lua.org/manual/5.3/manual.html#8.1

As CI now tests across versions, we need to change this value in the test to ensure compatibility with older Lua versions.

- Remove hardcoded version references in tests

Hardcoded version numbers in tests will mean that on a new version release, the tests using them will fail.

- create `constants.lua`to house lib-wide constants
- rename VERSION -> TRACKER_VERSION