Replace transaction_id with event_id (close #27) PR: #29
- replace tid with eid

tid (transaction id) is deprecated, and has been replaced with eid (event_id)

https://docs.snowplowanalytics.com/docs/collecting-data/collecting-from-own-applications/snowplow-tracker-protocol/#Event_transaction_parameters

- add spec/mocks/mock_uuid.lua

This allows us to use

package.loaded["lua_uuid"] = require("spec/mocks/mock_uuid")

to replace the entry in package.loaded, which Lua uses to control when and how modules are loaded, so when require("lua_uuid") is called in another file, it returns our mock_uuid.

The new package lua_uuid simply returns a single function when require'd, uuid(). By replacing this with our mock_uuid file, when uuid() is called in tracker.lua it runs the function in mock_uuid, which returns a known UUID.